Use btrblocks compressor
